Use PLCore 1.4.7 with Rewards parser
