+++ pom.xml
@@ -107,7 +107,7 @@
         <dependency>
             <groupId>pl.plajerlair</groupId>
             <artifactId>plcore</artifactId>
-            <version>1.4.6</version>
+            <version>1.4.7</version>
             <scope>compile</scope>
         </dependency>
         <dependency>
