+++ src/main/java/pl/plajer/villagedefense/arena/ArenaEvents.java
@@ -43,7 +43,7 @@ import pl.plajer.villagedefense.Main;
 import pl.plajer.villagedefense.api.StatsStorage;
 import pl.plajer.villagedefense.handlers.ChatManager;
 import pl.plajer.villagedefense.handlers.items.SpecialItemManager;
-import pl.plajer.villagedefense.handlers.reward.Reward;
+import pl.plajer.villagedefense.handlers.reward.GameReward;
 import pl.plajer.villagedefense.user.User;
 import pl.plajerlair.core.services.exception.ReportedException;
 import pl.plajerlair.core.utils.ItemBuilder;
@@ -126,7 +126,7 @@ public class ArenaEvents implements Listener {
             if (ArenaRegistry.getArena(e.getEntity().getKiller()) != null) {
               arena.addStat(e.getEntity().getKiller(), StatsStorage.StatisticType.KILLS);
               arena.addExperience(e.getEntity().getKiller(), 2);
-              plugin.getRewardsHandler().performReward(e.getEntity().getKiller(), Reward.RewardType.ZOMBIE_KILL);
+              plugin.getRewardsHandler().performReward(e.getEntity().getKiller(), GameReward.RewardType.ZOMBIE_KILL);
               plugin.getPowerupManager().spawnPowerup(e.getEntity().getLocation(), ArenaRegistry.getArena(e.getEntity().getKiller()));
             }
             return;
